Spring+React: Link User class to MySQL database
This involved creating a user repository, similar to an object manager in Django. With it, we can get an Iterable<User> of all users or a single User by its id. For some reason, the database doesn't seem to persist when the Spring server stops running, this will be looked into next.