Spring+React: Stop assigning the id of newly-created Users
The purpose of the @GeneratedValue annotation is to make the database generate the user's primary key, but we were breaking this by sending a key from the front-end and assigning it to the User in the backend. Nothing broke when the related lines of code were removed, so now we are actually generating primary keys for the users we create.